backend/requirements.txt
 
Flask>=2.0
flask-cors
pymongo
python-dotenv
werkzeug
